+++ tests/acceptance/CreateContentModelRelationFieldCest.php
@@ -1,4 +1,5 @@
 <?php
+use Codeception\Util\Locator;
 
 class CreateContentModelRelationFieldCest {
 
@@ -6,6 +7,13 @@ class CreateContentModelRelationFieldCest {
 		$i->maximizeWindow();
 		$i->loginAsAdmin();
 		$i->haveContentModel( 'Employee', 'Employees' );
+		$i->wait( 1 );
+		// Create a text field to check for reverse relationship conflicts with a reference model.
+		$i->click( 'Text', '.field-buttons' );
+		$i->fillField( [ 'name' => 'name' ], 'conflictsWithEmployeesSlug' );
+		$i->click( '.open-field button.primary' );
+		$i->wait( 1 );
+
 		$i->haveContentModel( 'Company', 'Companies' );
 		$i->wait( 1 );
 		$i->click( 'Relation', '.field-buttons' );
@@ -104,5 +112,53 @@ class CreateContentModelRelationFieldCest {
 		$i->click( '.open-field button.primary' );
 		$i->wait( 1 );
 		$i->see( 'Exceeds max length' );
+	}
+
+	public function i_can_see_warnings_when_a_reverse_relationship_slug_conflicts( AcceptanceTester $i ) {
+		// Attempt to create a relationship field with a reverse slug that
+		// conflicts with a field slug in the reference model.
+		$i->selectOption( '#reference', 'Employees' );
+		$i->click( '#enable-reverse' );
+		$i->wait( 1 );
+		$i->fillField( [ 'name' => 'reverseName' ], 'conflictsWithEmployeesSlug' );
+		$i->click( '.open-field button.primary' );
+		$i->wait( 1 );
+		$i->see( 'A field in the employee model has the same identifier' );
+
+		// Resolve the reverse slug conflict and save the field.
+		$i->fillField( [ 'name' => 'reverseName' ], 'resolveConflict' );
+		$i->click( '.open-field button.primary' );
+		$i->wait( 1 );
+		$i->see( 'Company Employees', '.field-list div.widest' ); // Field was created successfully.
+
+		// Attempt to create a second relationship field with the same reverse
+		// relationship slug as the relationship field we just saved.
+		$i->click( Locator::lastElement( '.add-item' ) );
+		$i->click( 'Relation', '.field-buttons' );
+		$i->wait( 1 );
+		$i->fillField( [ 'name' => 'name' ], 'Second Relationship' );
+		$i->selectOption( '#reference', 'Employees' );
+		$i->click( '#enable-reverse' );
+		$i->wait( 1 );
+		$i->fillField( [ 'name' => 'reverseName' ], 'resolveConflict' );
+		$i->click( '.open-field button.primary' );
+		$i->wait( 1 );
+		$i->see( 'A relationship field in this model has the same identifier.' );
+
+		// Resolve the reverse slug conflict.
+		$i->fillField( [ 'name' => 'reverseName' ], 'fixedConflict' );
+		$i->click( '.open-field button.primary' );
+		$i->wait( 1 );
+		$i->see( 'Second Relationship', Locator::lastElement( '.field-list div.widest' ) ); // Second field was created successfully.
+
+		// Open and update one of the relationship fields.
+		// To prevent reoccurence of an issue where reverse slugs could conflict
+		// with themselves and stop a relationship field from updating.
+		$i->clickWithLeftButton( '.field-list button.edit', -5, -5 );
+		$i->wait( 1 );
+		$i->fillField( [ 'name' => 'name' ], 'Updated Name' );
+		$i->click( '.open-field button.primary' );
+		$i->wait( 1 );
+		$i->see( 'Updated Name', '.field-list div.widest' );
 	}
 }