+++ tests/integration/api-validation/test-rest-model-endpoints.php
@@ -17,8 +17,8 @@ class RestModelEndpointTests extends WP_UnitTestCase {
 	private $route     = '/atlas/content-model';
 	private $test_models;
 
-	public function setUp(): void {
-		parent::setUp();
+	public function set_up(): void {
+		parent::set_up();
 
 		$this->test_models = $this->get_models();
 
@@ -44,8 +44,8 @@ class RestModelEndpointTests extends WP_UnitTestCase {
 		$this->post_ids = $this->get_post_ids();
 	}
 
-	public function tearDown() {
-		parent::tearDown();
+	public function tear_down() {
+		parent::tear_down();
 		wp_set_current_user( null );
 		global $wp_rest_server;
 		$wp_rest_server = null;