+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Atlas Content Modeler Changelog
 
+## 0.9.0 - 2021-11-03
+### Added
+- Fields now have an optional description which can be used to inform publishers what a field is for. The description is displayed above the field on the post entry screen.
+- Relationship field now supports one-to-one and one-to-many connections.
+
+### Fixed
+- ACM now prevents field slugs from conflicting with reserved slugs already used by WPGraphQL.
+- One-to-one field cardinality is now properly enforced.
+- API Identifier character limit is now properly enforced.
+- API Identifier validation now properly accounts for dashes and underscores.
+- Reverse API Identifier conflict detection no longer causes conflicts with the field being edited.
+- Added `publicly_queryable` parameter when registering custom post types. This fixes an issue caused by changes in [WPGraphQL #2134](https://github.com/wp-graphql/wp-graphql/issues/2134).
+
+### Changed
+- Posts without a title field no longer have their title saved as "Auto Draft". The fallback title is now entry[post id goes here].
+
 ## 0.8.0 - 2021-10-13
 
 ### Added